+++ flask/templates/index.html
@@ -1,0 +1,42 @@
+<!-- <!DOCTYPE html>
+<html>
+
+<head>
+  <title>Our Company</title>
+</head> -->
+
+<!-- <body>
+
+  <h1>제주도 날씨와 맛집</h1>
+  <h2></h2>
+
+  <p>위치를 입력해주세요!</p>
+  <form action="main.html">
+    <h2>
+      <input type="text" name="location" 위치="위치" placeholder="현재 위치">
+    </h2>
+    <button type="submit">입력</button>
+  </form>
+</body>
+</html> -->
+
+<!DOCTYPE html>
+        <html>
+        <head>
+        <title>제주도 날씨와 맛집 정보</title>
+        </head>
+        <body>
+        <h1>제주도 날씨와 맛집 정보🍊</h1>
+        <h2></h2>
+        <p>날짜를 입력해주세요🗓️</p>
+        <form action="/cgi-bin/process_form.py" method="post">
+            <h2>
+            <label for="start_date">Start Date:</label>
+            <input type="date" id="start_date" name="start_date" required>
+            <label for="end_date">End Date:</label>
+            <input type="date" id="end_date" name="end_date" required>
+            </h2>
+            <button type="submit">Enter</button>
+        </form>
+        </body>
+        </html>
