+++ flask/templates/page3.html
@@ -1,0 +1,193 @@
+<!DOCTYPE html>
+<html>
+   <head>
+      <title> 잘 도착했지롱 🌈 </title>
+      <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+      <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
+      <style>
+         body {
+         display: flex;
+         /* Flex 컨테이너로 설정 */
+         flex-direction: column;
+         justify-content: center;
+         /* 가로 가운데 정렬 */
+         align-items: center;
+         /* 세로 가운데 정렬 */
+         background-image: url("{{ url_for('static', filename='background.png') }}");
+         background-size: cover;
+         /* 배경 이미지를 화면에 맞게 조절 */
+         background-repeat: no-repeat;
+         /* 배경 이미지를 반복하지 않음 */
+         background-position: center;
+         /* 배경 이미지를 가운데로 정렬 */
+         background-attachment: fixed;
+         }
+         .matrix {
+         display: block;
+         margin-bottom: 20px;
+         }
+         .category {
+         background-color: rgba(255, 255, 255, 0.7);
+         width: 100%;
+         padding: 10px;
+         border: 1px solid #ccc;
+         }
+         .subcategory {
+         width: 100%;
+         padding: 10px;
+         border: 1px solid #ccc;
+         }
+         tbody {
+         border: 2px solid #dee2e6;
+         }
+         /* 테이블 머리글의 테두리 두께를 2px로 조절 */
+         thead {
+         border: 2px solid #dee2e6;
+         }
+      </style>
+   </head>
+   <body>
+      <h1> 잘 도착했지롱 🌈 </h1>
+      <h3>여행 기간: {{ start_date }} ~ {{ end_date }} </h3>
+      <h3> 여행 위치: {{ region }} </h3>
+      <div class="matrix">
+         <div class="category">
+            <h2> 여행 경비 정보 </h2>
+            <h3 class="text-center"> 총 금액: {{ total_price }} </h3>
+            <div class="subcategory">
+               <h3>항공권</h3>
+               <div class="subcategory" , name="항공권_to">
+                  {% if selected_output['항공권_to']|length > 0 %} 
+                  <h4>가는 편</h4>
+                  <table class="table table-bordered text-center">
+                     <thead class="table-light">
+                        <tr>
+                           {% for key in selected_output['항공권_to'].keys() %} 
+                           <th name="{{ key }}">{{ key }}</th>
+                           {% endfor %} 
+                        </tr>
+                     </thead>
+                     <tbody>
+                        <tr>
+                           {% for key, value in selected_output['항공권_to'].items() %} 
+                           <td name="{{ key }}">{{ value }}</td>
+                           {% endfor %} 
+                        </tr>
+                     </tbody>
+                  </table>
+                  {% endif %}
+               </div>
+               <div class="subcategory" , name="항공권_from">
+                  {% if selected_output['항공권_from']|length > 0 %} 
+                  <h4>오는 편</h4>
+                  <table class="table table-bordered text-center">
+                     <thead class="table-light">
+                        <tr>
+                           {% for key in selected_output['항공권_from'].keys() %} 
+                           <th name="{{ key }}">{{ key }}</th>
+                           {% endfor %} 
+                        </tr>
+                     </thead>
+                     <tbody>
+                        <tr>
+                           {% for key, value in selected_output['항공권_from'].items() %} 
+                           <td name="{{ key }}">{{ value }}</td>
+                           {% endfor %} 
+                        </tr>
+                     </tbody>
+                  </table>
+                  {% endif %}
+               </div>
+            </div>
+            <div class="subcategory" , name="호텔">
+               <h3>숙박시설</h3>
+               {% if selected_output['호텔']|length > 0 %} 
+               <table class="table table-bordered text-center">
+                  <thead class="table-light">
+                     <tr>
+                        {% for key in selected_output['호텔'].keys() %} 
+                        <th name="{{ key }}">{{ key }}</th>
+                        {% endfor %} 
+                     </tr>
+                  </thead>
+                  <tbody>
+                     <tr>
+                        {% for key, value in selected_output['호텔'].items() %} 
+                        <td name="{{ key }}">{{ value }}</td>
+                        {% endfor %} 
+                     </tr>
+                  </tbody>
+               </table>
+               {% endif %}
+            </div>
+            <div class="subcategory" , name="렌트카">
+               <h3>렌터카</h3>
+               {% if selected_output['렌트카']|length > 0 %} 
+               <table class="table table-bordered text-center">
+                  <thead class="table-light">
+                     <tr>
+                        {% for key in selected_output['렌트카'].keys() %} 
+                        <th name="{{ key }}">{{ key }}</th>
+                        {% endfor %} 
+                     </tr>
+                  </thead>
+                  <tbody>
+                     <tr>
+                        {% for key, value in selected_output['렌트카'].items() %} 
+                        <td name="{{ key }}">{{ value }}</td>
+                        {% endfor %} 
+                     </tr>
+                  </tbody>
+               </table>
+               {% endif %}
+            </div>
+         </div>
+         <div class="category">
+            <h2> 제주도 날씨 정보🍊 </h2>
+            {% if weather_output|length > 0 %} 
+            <table class="table table-bordered text-center">
+                <thead class="table-light">
+                    {% for key in weather_output[0] %}
+                    <th name="{{ key }}">{{ key }}</th>
+                    {% endfor %}
+                </thead>
+                <tbody>
+                    {% for values in weather_output.values() %}
+                    <tr>
+                        {% for key, value in values.items() %}
+                        {% if key in ['Day', 'Weather'] %}
+                        <td name="{{ key }}">{{ value }}</td>
+                        {% else %}
+                        <td name="{{ key }}">{{ value }}°C</td>
+                        {% endif %}
+                        {% endfor %}
+                    </tr>
+                    {% endfor %}
+                </tbody>
+            </table>
+            {% endif %}
+         </div>
+         <div class="category">
+            <h2> 제주도 맛집 정보🍊 </h2>
+            {% if mangoplate_output|length > 0 %} 
+            <table class="table table-bordered text-center">
+                <thead class="table-light">
+                    {% for key in mangoplate_output[0] %}
+                    <th name="{{ key }}">{{ key }}</th>
+                    {% endfor %}
+                </thead>
+                <tbody>
+                    {% for values in mangoplate_output.values() %}
+                    <tr>
+                        {% for key, value in values.items() %}
+                        <td name="{{ key }}">{{ value }}</td>
+                        {% endfor %}
+                    </tr>
+                    {% endfor %}
+                </tbody>
+            </table>
+            {% endif %}
+         </div>
+      </div>
+   </body>
+</html>
