+++ flask/weather_DB.csv
@@ -1,0 +1,32 @@
+﻿Day,Weather,Low Temperature,High Temperature
+1,흐리고 습하며 때때로 비와 뇌우가 내림,24,31
+2,약간 흐리고 오후 중 때때로 소나기가 내림,24,29
+3,대체로 흐리고 때때로 비와 뇌우가 내림,24,30
+4,흐리고 오후 중 일부 지역에 뇌우가 내림,24,29
+5,약간 흐리고 소나기나 뇌우가 내림,26,30
+6,대체로 흐리고 오후 중 때때로 소나기가 내림,25,31
+7,대체로 흐림,24,30
+8,약간 흐림,25,31
+9,약간 흐리고 습함,25,30
+10,대체로 맑고 습함,26,31
+11,맑고 습함,26,31
+12,맑고 습함,25,31
+13,대체로 맑고 습함,26,31
+14,맑고 습함,26,31
+15,오전 중 몇 차례 소나기; 대체로 흐리고 습함,25,30
+16,습하고 점차 흐려짐,26,31
+17,흐리고 습함; 오후에 몇 차례 소나기와 뇌우가 내림,26,31
+18,흐리고 바람이 많이 불며 몇 차례 소나기와 뇌우가 내림,25,30
+19,대체로 흐림,25,30
+20,약간 흐림,24,29
+21,점차 흐려짐,25,30
+22,흐리고 오후에 몇 차례 소나기와 뇌우가 내림,25,29
+23,흐리고 습함; 오후에 소나기와 뇌우가 내림,24,28
+24,약간 흐림,21,28
+25,맑음,22,28
+26,약간 흐림,22,29
+27,흐리고 오후에 가벼운 비가 내림,22,28
+28,대체로 흐리고 바람이 많이 불며 소나기가 내림,23,28
+29,흐리고 바람이 많이 불며 몇 차례 소나기가 내림,23,29
+30,대체로 흐리고 습하며 몇 차례 소나기가 내림,22,29
+31,흐리고 습하며 소나기와 뇌우가 내림,22,28
